Revert "minify cloudfront viewer request handler code to free up space for user injection"
This reverts commit b231b9b03d1ab4734a669103dcd6c39e0563475f.
