+++ platform/src/components/aws/ssr-site.ts
@@ -36,7 +36,7 @@ import { useProvider } from "./helpers/provider.js";
 import { Link } from "../link.js";
 import { URL_UNAVAILABLE } from "./linkable.js";
 import {
-  buildCfHandlerCode,
+  CF_ROUTER_INJECTION,
   CF_BLOCK_CLOUDFRONT_URL_INJECTION,
   KV_SITE_METADATA,
   ProtectionConfig,
@@ -882,10 +882,25 @@ export abstract class SsrSite extends Component implements Link.Linkable {
           {
             runtime: "cloudfront-js-2.0",
             keyValueStoreAssociations: kvStoreArn ? [kvStoreArn] : [],
-            code: buildCfHandlerCode("site-handler", {
-              kvNamespace,
-              injection: [userInjection, blockCloudfrontUrlInjection],
-            }),
+            code: interpolate`
+import cf from "cloudfront";
+async function handler(event) {
+  ${userInjection}
+  ${blockCloudfrontUrlInjection}
+  ${CF_ROUTER_INJECTION}
+
+  const kvNamespace = "${kvNamespace}";
+
+  // Load metadata
+  let metadata;
+  try {
+    const v = await cf.kvs().get(kvNamespace + ":metadata");
+    metadata = JSON.parse(v);
+  } catch (e) {}
+
+  await routeSite(kvNamespace, metadata);
+  return event.request;
+}`,
           },
           { parent: self },
         );
